moving type out
